refactor: split config into database, auth, email and widget modules
