+++ backend/app/config/dashboard.py
@@ -1,0 +1,9 @@
+"""Dashboard widget result limits"""
+
+# How many of the user's most recent transactions the dashboard's recent-activity widget returns.
+DASHBOARD_RECENT_TRANSACTIONS_LIMIT = 15
+
+# How many calendar months the dashboard's savings rate history returns, inclusive of the
+# current (in-progress) month. With the default of 7 the series covers the current month
+# plus the six prior months.
+DASHBOARD_SAVINGS_HISTORY_MONTHS = 7
