+++ backend/app/services/email/rendering.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
-from app.config import EMAIL_SENDER_NAME
+from app.config.email import EMAIL_SENDER_NAME
 from app.services.email.contract import RenderedEmail
 
 _TEMPLATES_DIR = Path(__file__).parent / "templates"
