+++ backend/tests/test_config_update_checks.py
@@ -1,6 +1,7 @@
 import pytest
 
-from app.config import _optional_bool_env, is_update_check_enabled
+from app.config.env import optional_bool_env
+from app.config.runtime import is_update_check_enabled
 
 
 @pytest.mark.parametrize("value", ["true", "TRUE"])
@@ -8,7 +9,7 @@ def test_optional_bool_env_parses_true_values(monkeypatch, value):
     """Boolean env parsing accepts true values"""
     monkeypatch.setenv("UPDATE_CHECKS_ENABLED", value)
 
-    assert _optional_bool_env("UPDATE_CHECKS_ENABLED", default=False) is True
+    assert optional_bool_env("UPDATE_CHECKS_ENABLED", default=False) is True
 
 
 @pytest.mark.parametrize("value", ["false", "FALSE"])
@@ -16,21 +17,21 @@ def test_optional_bool_env_parses_false_values(monkeypatch, value):
     """Boolean env parsing accepts false values"""
     monkeypatch.setenv("UPDATE_CHECKS_ENABLED", value)
 
-    assert _optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is False
+    assert optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is False
 
 
 def test_optional_bool_env_uses_default_for_missing_value(monkeypatch):
     """Boolean env parsing uses the default when missing"""
     monkeypatch.delenv("UPDATE_CHECKS_ENABLED", raising=False)
 
-    assert _optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is True
+    assert optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is True
 
 
 def test_optional_bool_env_uses_default_for_blank_value(monkeypatch):
     """Boolean env parsing uses the default when blank"""
     monkeypatch.setenv("UPDATE_CHECKS_ENABLED", " ")
 
-    assert _optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is True
+    assert optional_bool_env("UPDATE_CHECKS_ENABLED", default=True) is True
 
 
 @pytest.mark.parametrize("value", ["1", "0", "yes", "no", "on", "off", "maybe"])
@@ -39,7 +40,7 @@ def test_optional_bool_env_rejects_invalid_value(monkeypatch, value):
     monkeypatch.setenv("UPDATE_CHECKS_ENABLED", value)
 
     with pytest.raises(RuntimeError, match="Invalid UPDATE_CHECKS_ENABLED"):
-        _optional_bool_env("UPDATE_CHECKS_ENABLED", default=False)
+        optional_bool_env("UPDATE_CHECKS_ENABLED", default=False)
 
 
 @pytest.mark.parametrize(
